Refactor RdsFileUploader component and add supporting components for improved file upload functionality
